zynq: nand: Fix "unsupported start_cmd" issue.
The zynq nand driver will report unsupported start cmd due to
the u8 size used in start_cmd and end_cmd variable in zynq_nand.c
file was not sufficient to handle nand command such as
NAND_CMD_READ0_ONLY. Hence, the zynq nand driver will reporting
"unsupported start command". This commit is to change the start_cmd
and end_cmd variable size from u8 to int.

+++ drivers/mtd/nand/zynq_nand.c
@@ -148,8 +148,8 @@ struct zynq_nand_info {
  * @end_cmd_valid:	The second cycle command is valid for cmd or data phase
  */
 struct zynq_nand_command_format {
-	u8 start_cmd;
-	u8 end_cmd;
+	int start_cmd;
+	int end_cmd;
 	u8 addr_cycles;
 	u8 end_cmd_valid;
 };
